+++ sketch-core/src/androidMain/kotlin/com/github/panpf/sketch/source/ContentDataSource.kt
@@ -18,7 +18,6 @@ package com.github.panpf.sketch.source
 
 import android.content.Context
 import android.net.Uri
-import androidx.annotation.WorkerThread
 import com.github.panpf.sketch.Sketch
 import com.github.panpf.sketch.fetch.isFileUri
 import com.github.panpf.sketch.source.DataFrom.LOCAL
@@ -42,21 +41,19 @@ class ContentDataSource constructor(
 
     override val dataFrom: DataFrom = LOCAL
 
-    @WorkerThread
     @Throws(IOException::class)
-    override fun openSourceOrNull(): Source =
+    override fun openSource(): Source =
         (context.contentResolver.openInputStream(contentUri)
             ?: throw IOException("Invalid content uri: $contentUri")).source()
 
-    @WorkerThread
     @Throws(IOException::class)
-    override fun getFileOrNull(sketch: Sketch): Path? {
+    override fun getFile(sketch: Sketch): Path {
         val sketchUri = contentUri.toString().toUri()
         return if (isFileUri(sketchUri)) {
             val filePath = sketchUri.path!!
             File(filePath).toOkioPath()
         } else {
-            getDataSourceCacheFile(sketch, this)
+            cacheFile(sketch)
         }
     }