+++ sketch-core/src/androidMain/kotlin/com/github/panpf/sketch/source/ResourceDataSource.kt
@@ -19,7 +19,6 @@ package com.github.panpf.sketch.source
 import android.content.res.Resources
 import androidx.annotation.DrawableRes
 import androidx.annotation.RawRes
-import androidx.annotation.WorkerThread
 import com.github.panpf.sketch.Sketch
 import com.github.panpf.sketch.fetch.newResourceUri
 import com.github.panpf.sketch.source.DataFrom.LOCAL
@@ -41,13 +40,11 @@ class ResourceDataSource constructor(
 
     override val dataFrom: DataFrom = LOCAL
 
-    @WorkerThread
     @Throws(IOException::class)
-    override fun openSourceOrNull(): Source = resources.openRawResource(resId).source()
+    override fun openSource(): Source = resources.openRawResource(resId).source()
 
-    @WorkerThread
     @Throws(IOException::class)
-    override fun getFileOrNull(sketch: Sketch): Path? = getDataSourceCacheFile(sketch, this)
+    override fun getFile(sketch: Sketch): Path = cacheFile(sketch)
 
     override fun equals(other: Any?): Boolean {
         if (this === other) return true